🏗️ refactor: modularize workflow architecture into domain-specific modules
- Split monolithic workflows.ts into 6 specialized modules:
  - workflows/generate.ts: Template generation workflows
  - workflows/git.ts: Git branch and commit workflows
  - workflows/init.ts: Project initialization workflows
  - workflows/release.ts: Version release workflows
  - workflows/validate.ts: Validation workflows
  - workflows/index.ts: Centralized exports
- Update index.ts to re-export from new workflow modules
- Maintain backward compatibility with existing API
- Improve maintainability and separation of concerns
