--- .gitignore
@@ -1,40 +1,0 @@
-cat <<EOF > .gitignore
-# Dependencies
-node_modules/
-.pnp
-.pnp.js
-.yarn/install-state.gz
-
-# Production
-build/
-dist/
-.out/
-
-# Misc
-.DS_Store
-*.pem
-
-# Logs
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-lerna-debug.log*
-
-# Environment Variables
-.env
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
-# IDEs and editors
-.vscode/*
-!.vscode/settings.json
-!.vscode/tasks.json
-!.vscode/launch.json
-!.vscode/extensions.json
-*.sublime-workspace
-
-# TypeScript
-*.tsbuildinfo
-EOFpackages/vscode-extension/coverage/
